Refactor: Separate Logic, Mocks, and Tests
